conversao_frameworks/requirements.txt
pytest-playwright
selenium
caqui